feat: add request / response durations & counts by operation (#173)
