+++ src/main/java/io/gravitee/reporter/elasticsearch/mapping/es7/ES7IndexPreparer.java
@@ -17,15 +17,9 @@ package io.gravitee.reporter.elasticsearch.mapping.es7;
 
 import io.gravitee.common.templating.FreeMarkerComponent;
 import io.gravitee.elasticsearch.client.Client;
-import io.gravitee.elasticsearch.utils.Type;
 import io.gravitee.reporter.elasticsearch.config.PipelineConfiguration;
 import io.gravitee.reporter.elasticsearch.config.ReporterConfiguration;
 import io.gravitee.reporter.elasticsearch.mapping.AbstractIndexPreparer;
-import io.reactivex.rxjava3.core.Completable;
-import io.reactivex.rxjava3.core.CompletableSource;
-import io.reactivex.rxjava3.functions.Function;
-import java.util.Collections;
-import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 
 /**
@@ -41,46 +35,6 @@ public class ES7IndexPreparer extends AbstractIndexPreparer {
         final FreeMarkerComponent freeMarkerComponent,
         final Client client
     ) {
-        super(configuration, pipelineConfiguration, freeMarkerComponent, client);
-    }
-
-    @Override
-    public Completable prepare() {
-        return indexMapping().andThen(pipeline());
-    }
-
-    @Override
-    protected Function<Type, CompletableSource> indexTypeMapper() {
-        return type -> {
-            final String typeName = type.getType();
-            boolean dataStream = type.isDataStream();
-            final String templateName = configuration.getIndexName() + '-' + typeName;
-            final String aliasName = configuration.getIndexName() + '-' + typeName;
-
-            log.debug("Trying to put template mapping for type[{}] name[{}]", typeName, templateName);
-
-            Map<String, Object> data = getTemplateData();
-            data.put("indexName", configuration.getIndexName() + '-' + typeName);
-
-            final String template = freeMarkerComponent.generateFromTemplate("/es7x/mapping/index-template-" + typeName + ".ftl", data);
-            final Completable templateCreationCompletable = client.putTemplate(templateName, template);
-
-            if (configuration.isIlmManagedIndex() && !dataStream) {
-                return templateCreationCompletable.andThen(ensureAlias(aliasName));
-            }
-            return templateCreationCompletable;
-        };
-    }
-
-    private Completable ensureAlias(String aliasName) {
-        final String aliasTemplate = freeMarkerComponent.generateFromTemplate(
-            "/es7x/alias/alias.ftl",
-            Collections.singletonMap("aliasName", aliasName)
-        );
-
-        return client
-            .getAlias(aliasName)
-            .switchIfEmpty(client.createIndexWithAlias(aliasName + "-000001", aliasTemplate).toMaybe())
-            .ignoreElement();
+        super(configuration, pipelineConfiguration, freeMarkerComponent, client, "/esx7");
     }
 }
