--- src/main/resources/freemarker/es8x/alias/alias.ftl
@@ -1,9 +1,0 @@
-<@compress single_line=true>
-{
-    "aliases": {
-        "${aliasName}": {
-            "is_write_index": true
-        }
-    }
-}
-</@compress>
